Move MemoryManager into new o.a.s.storage.memory package
